refactor/move login to custom hooks
