+++ src/hooks/usePaginationParams.ts
@@ -1,0 +1,17 @@
+import { useSearchParams } from "next/navigation";
+import { getPageParam } from "@/utils/getPageParam";
+import { getLimitParam } from "@/utils/getLimitParam";
+
+export const usePaginationParams = () => {
+  const searchParams = useSearchParams();
+
+  const limitParam = searchParams.get("limit");
+
+  const page = getPageParam(searchParams.get("page"));
+  const limit = getLimitParam(isNaN(Number(limitParam)) ? "3" : limitParam);
+
+  return {
+    page,
+    limit,
+  };
+};
